Move SessionStore to sessions/core/
