+++ packages/core/src/sessions/core/create-session.spec.ts
@@ -5,8 +5,8 @@ import { notStrictEqual, ok, strictEqual } from 'assert';
 import { createService } from '../../core';
 import { createSession } from './create-session';
 import { Session } from './session';
-import { SessionState } from '../session-state.interface';
-import { SessionStore } from '../session-store';
+import { SessionState } from './session-state.interface';
+import { SessionStore } from './session-store';
 
 describe('createSession', () => {
