+++ packages/core/src/sessions/core/session-state.interface.ts
